requirements.txt
argparse==1.2.1
hurry.filesize==0.9
python-dateutil==2.2
requests==2.0.1
six==1.4.1
wsgiref==0.1.2
